move cert generation into module
